update folder and file structure
