+++ package.json
@@ -21,7 +21,6 @@
     "@mui/x-date-pickers": "^5.0.8",
     "date-fns": "^2.29.3",
     "mongoose": "^6.7.2",
-    "nanoid": "^4.0.0",
     "next": "13.0.3",
     "next-router-mock": "^0.8.0",
     "react": "18.2.0",
