+++ pages/index.js
@@ -2,7 +2,7 @@ import useSWR from "swr";
 import EventCard from "../components/EventCard";
 import StyledLink from "../components/StyledLink";
 import Svg from "../components/Svg";
-import { fetcher } from "../helpers/api";
+import { fetcher } from "../utils/api";
 
 export default function Home() {
   const { data: events, error } = useSWR("/api/events", fetcher);
