+++ services/eventService.js
@@ -1,23 +1,5 @@
-import mongoose, { Schema, model, models } from "mongoose";
-import { nanoid } from "nanoid";
-
-const URI = `mongodb+srv://gaiisay:${process.env.MONGODB_PASSWORD}@cluster0.nhyoetz.mongodb.net/?retryWrites=true&w=majority`;
-
-const eventSchema = new Schema({
-  id: String,
-  name: String,
-  description: String,
-  date: String,
-  startTime: String,
-  endTime: String,
-  location: String,
-});
-
-const Event = models.Event || model("Event", eventSchema);
-
-async function connectToDatabase() {
-  await mongoose.connect(URI);
-}
+import Event from "../models/Event";
+import connectToDatabase from "../utils/dbConnect";
 
 async function getAllEvents() {
   await connectToDatabase();
@@ -55,7 +37,7 @@ async function createEvent(event) {
 
   const createdEvent = await Event.create({
     ...event,
-    id: nanoid(),
+    id: crypto.randomUUID(),
   });
 
   return {
